Reorganized files into folders
