+++ duelinvaders/src/main/java/org/enstabretagne/Core/Canon.java
@@ -1,4 +1,4 @@
-package org.enstabretagne;
+package org.enstabretagne.Core;
 
 import java.util.Date;
