+++ duelinvaders/src/main/java/org/enstabretagne/Core/MoteurJeu.java
@@ -1,4 +1,4 @@
-package org.enstabretagne;
+package org.enstabretagne.Core;
 
 import java.util.ArrayList;
