+++ duelinvaders/src/main/java/org/enstabretagne/Game/GameLauncher.java
@@ -1,4 +1,4 @@
-package org.enstabretagne;
+package org.enstabretagne.Game;
 
 import com.almasb.fxgl.app.GameApplication;
 import com.almasb.fxgl.app.GameSettings;
